refactor: blogNav
